Simple fix for YTAS problems
These are naive, partial fixes for the YTAS issues with `i`, `j` and `f`. For the `i`-dot, it would be better to have a special bold master that compresses the dots, and they would benefit from being components. But this simply does not lower the `i`-dots so much, and lowers the `f`-bar a bit.

- https://github.com/googlefonts/roboto-flex/issues/319
- https://github.com/googlefonts/roboto-flex/issues/314